requirements.txt
numpy==1.24.3
cupy-cuda12x  # Replace '12x' with your specific CUDA version